improvements for input format for api calls, tested with and without input format
