--- sim/app/w/[id]/components/workflow-block/components/sub-block/components/input-format.tsx
@@ -1,183 +1,0 @@
-import { Plus, Trash } from 'lucide-react'
-import { ChevronDown } from 'lucide-react'
-import { Badge } from '@/components/ui/badge'
-import { Button } from '@/components/ui/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { cn } from '@/lib/utils'
-import { useSubBlockValue } from '../hooks/use-sub-block-value'
-
-interface InputField {
-  id: string
-  name: string
-  type: 'string' | 'number' | 'boolean' | 'object' | 'array'
-  collapsed?: boolean
-}
-
-interface InputFormatProps {
-  blockId: string
-  subBlockId: string
-}
-
-// Default values
-const DEFAULT_FIELD: InputField = {
-  id: crypto.randomUUID(),
-  name: '',
-  type: 'string',
-  collapsed: false,
-}
-
-export function InputFormat({ blockId, subBlockId }: InputFormatProps) {
-  // State hooks
-  const [value, setValue] = useSubBlockValue<InputField[]>(blockId, subBlockId)
-  const fields = value || [DEFAULT_FIELD]
-
-  // Field operations
-  const addField = () => {
-    const newField: InputField = {
-      ...DEFAULT_FIELD,
-      id: crypto.randomUUID(),
-    }
-    setValue([...fields, newField])
-  }
-
-  const removeField = (id: string) => {
-    if (fields.length === 1) return
-    setValue(fields.filter((field) => field.id !== id))
-  }
-
-  // Update handlers
-  const updateField = (id: string, field: keyof InputField, value: any) => {
-    setValue(fields.map((f) => (f.id === id ? { ...f, [field]: value } : f)))
-  }
-
-  const toggleCollapse = (id: string) => {
-    setValue(fields.map((f) => (f.id === id ? { ...f, collapsed: !f.collapsed } : f)))
-  }
-
-  // Field header
-  const renderFieldHeader = (field: InputField, index: number) => (
-    <div
-      className="flex h-9 items-center justify-between px-3 py-1 cursor-pointer"
-      onClick={() => toggleCollapse(field.id)}
-    >
-      <div className="flex items-center">
-        <span className="text-sm text-foreground">
-          {field.name ? field.name : `Field ${index + 1}`}
-        </span>
-        {field.name && (
-          <Badge variant="outline" className="ml-2 py-0 h-5 text-xs font-normal bg-muted">
-            {field.type}
-          </Badge>
-        )}
-      </div>
-      <div className="flex items-center gap-1" onClick={(e) => e.stopPropagation()}>
-        <Button variant="ghost" size="icon" onClick={addField} className="h-6 w-6 rounded-full">
-          <Plus className="h-3.5 w-3.5" />
-          <span className="sr-only">Add Field</span>
-        </Button>
-
-        <Button
-          variant="ghost"
-          size="icon"
-          onClick={() => removeField(field.id)}
-          disabled={fields.length === 1}
-          className="h-6 w-6 rounded-full text-destructive hover:text-destructive"
-        >
-          <Trash className="h-3.5 w-3.5" />
-          <span className="sr-only">Delete Field</span>
-        </Button>
-      </div>
-    </div>
-  )
-
-  // Main render
-  return (
-    <div className="space-y-2">
-      {fields.map((field, index) => (
-        <div
-          key={field.id}
-          data-field-id={field.id}
-          className={cn(
-            'rounded-md border border-input bg-transparent shadow-sm',
-            field.collapsed ? 'overflow-hidden' : 'overflow-visible'
-          )}
-        >
-          {renderFieldHeader(field, index)}
-
-          {!field.collapsed && (
-            <div className="px-3 pt-1.5 pb-2 space-y-2 border-t">
-              <div className="space-y-1.5">
-                <Label className="text-xs">Name</Label>
-                <Input
-                  name="name"
-                  value={field.name}
-                  onChange={(e) => updateField(field.id, 'name', e.target.value)}
-                  placeholder="firstName"
-                  className="h-9 placeholder:text-muted-foreground/50"
-                />
-              </div>
-
-              <div className="space-y-1.5">
-                <Label className="text-xs">Type</Label>
-                <DropdownMenu>
-                  <DropdownMenuTrigger asChild>
-                    <Button variant="outline" className="w-full justify-between h-9 font-normal">
-                      <div className="flex items-center">
-                        <span>{field.type}</span>
-                      </div>
-                      <ChevronDown className="h-4 w-4 opacity-50" />
-                    </Button>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent align="end" className="w-[200px]">
-                    <DropdownMenuItem
-                      onClick={() => updateField(field.id, 'type', 'string')}
-                      className="cursor-pointer"
-                    >
-                      <span className="font-mono mr-2">Aa</span>
-                      <span>String</span>
-                    </DropdownMenuItem>
-                    <DropdownMenuItem
-                      onClick={() => updateField(field.id, 'type', 'number')}
-                      className="cursor-pointer"
-                    >
-                      <span className="font-mono mr-2">123</span>
-                      <span>Number</span>
-                    </DropdownMenuItem>
-                    <DropdownMenuItem
-                      onClick={() => updateField(field.id, 'type', 'boolean')}
-                      className="cursor-pointer"
-                    >
-                      <span className="font-mono mr-2">0/1</span>
-                      <span>Boolean</span>
-                    </DropdownMenuItem>
-                    <DropdownMenuItem
-                      onClick={() => updateField(field.id, 'type', 'object')}
-                      className="cursor-pointer"
-                    >
-                      <span className="font-mono mr-2">{'{}'}</span>
-                      <span>Object</span>
-                    </DropdownMenuItem>
-                    <DropdownMenuItem
-                      onClick={() => updateField(field.id, 'type', 'array')}
-                      className="cursor-pointer"
-                    >
-                      <span className="font-mono mr-2">[]</span>
-                      <span>Array</span>
-                    </DropdownMenuItem>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              </div>
-            </div>
-          )}
-        </div>
-      ))}
-    </div>
-  )
-}
