+++ sim/app/w/[id]/components/workflow-block/components/sub-block/sub-block.tsx
@@ -13,11 +13,11 @@ import { Dropdown } from './components/dropdown'
 import { EvalInput } from './components/eval-input'
 import { FileSelectorInput } from './components/file-selector/file-selector-input'
 import { FolderSelectorInput } from './components/folder-selector/components/folder-selector-input'
-import { InputFormat } from './components/input-format'
 import { LongInput } from './components/long-input'
 import { ScheduleConfig } from './components/schedule/schedule-config'
 import { ShortInput } from './components/short-input'
 import { SliderInput } from './components/slider-input'
+import { InputFormat } from './components/starter/input-format'
 import { Switch } from './components/switch'
 import { Table } from './components/table'
 import { TimeInput } from './components/time-input'
